+++ app/src/main/java/space/jiyun/coala/main/MainNavigator.kt
@@ -1,4 +1,4 @@
-package space.jiyun.coala
+package space.jiyun.coala.main
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
