+++ app/src/main/java/space/jiyun/coala/main/MainViewModel.kt
@@ -1,15 +1,11 @@
-package space.jiyun.coala
+package space.jiyun.coala.main
 
 import android.view.MenuItem
-import androidx.annotation.IdRes
-import androidx.databinding.BindingAdapter
 import androidx.databinding.BindingMethod
 import androidx.databinding.BindingMethods
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.ViewModel
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import space.jiyun.coala.databinding.AnswerFragmentBinding
+import space.jiyun.coala.R
 import space.jiyun.coala.ui.answer.AnswerFragment
 import space.jiyun.coala.ui.home.HomeFragment
 import space.jiyun.coala.ui.question.QuestionFragment
