+++ app/src/main/res/layout/main_activity.xml
@@ -8,7 +8,7 @@
 
         <variable
             name = "viewmodel"
-            type = "space.jiyun.coala.MainViewModel"
+            type = "space.jiyun.coala.main.MainViewModel"
             />
     </data>
 
@@ -16,7 +16,7 @@
         android:id = "@+id/container"
         android:layout_width = "match_parent"
         android:layout_height = "match_parent"
-        tools:context = ".MainActivity"
+        tools:context = ".main.MainActivity"
         >
 
         <androidx.appcompat.widget.Toolbar
@@ -26,7 +26,6 @@
             android:background = "@android:color/white"
             android:minHeight = "?attr/actionBarSize"
             app:layout_constraintTop_toTopOf = "parent"
-            android:theme="@style/Toolbar"
             app:popupTheme = "@style/ThemeOverlay.AppCompat.Light"
             />
